romfs: moved the standalone extractor cli to the handler
The file is reorganized in a followup commit
